# apps/ai-orchestrator/openapi/orchestrator-openapi.yaml
openapi: 3.1.0
info:
  title: AI Orchestrator API
  version: 0.1.0
  description: >
    Minimal orchestration API for generating YAML test cases via the test-design-agent
    and optionally dispatching the web-playwright-python runner.
servers:
  - url: http://127.0.0.1:8000
paths:
...
  /assets/scaffold/templates:
    get:
      summary: List available scaffold templates
      operationId: getScaffoldTemplates
      responses:
        "200":
          description: Available scaffold templates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScaffoldTemplateListResponse"
        "500":
          description: Internal template discovery failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /assets/scaffold/templates/{template}:
    get:
      summary: Get a scaffold template definition
      operationId: getScaffoldTemplate
      parameters:
        - in: path
          name: template
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scaffold template details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScaffoldTemplateResponse"
        "422":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal template discovery failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reports/latest:
    get:
      summary: Get the latest execution report
      operationId: getLatestExecutionReport
      responses:
        "200":
          description: Latest execution report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionReportResponse"
        "422":
          description: Report lookup failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal report lookup failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reports/{case_id}:
    get:
      summary: Get an execution report by case id
      operationId: getExecutionReportByCaseId
      parameters:
        - in: path
          name: case_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Execution report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionReportResponse"
        "422":
          description: Report lookup failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal report lookup failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /runners/catalog:
    get:
      summary: List supported runner profiles
      operationId: getRunnerCatalog
      responses:
        "200":
          description: Runner catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunnerCatalogResponse"
        "500":
          description: Internal runner catalog failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orchestrate:
    post:
      summary: Generate a test case and optionally execute it
      operationId: postOrchestrate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrchestrateRequest"
      responses:
        "201":
          description: Test case generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrchestrationResult"
        "400":
          description: Invalid JSON payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Route not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Runner execution failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal orchestration failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requirements/parse:
    post:
      summary: Parse requirement into structured test intents
      operationId: postParseRequirement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseRequirementRequest"
      responses:
        "201":
          description: Requirement parsed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseRequirementResponse"
        "400":
          description: Invalid JSON payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Route not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal requirement parsing failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requirements/telemetry/summary:
    get:
      summary: Summarize requirement parse telemetry
      operationId: getRequirementParseTelemetrySummary
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 500
        - in: query
          name: prompt_version
          required: false
          schema:
            type: string
        - in: query
          name: model
          required: false
          schema:
            type: string
        - in: query
          name: mode
          required: false
          schema:
            type: string
        - in: query
          name: stage
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Telemetry summary
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /risk/evaluate:
    post:
      summary: Evaluate risk for a requirement and execution plan
      operationId: postRiskEvaluate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                requirement_spec:
                  $ref: "#/components/schemas/RequirementSpecV1"
                execution_plan:
                  type: object
                  additionalProperties: true
                execution_record:
                  $ref: "#/components/schemas/ExecutionRecordV1"
                failure_analysis:
                  type: object
                  additionalProperties: true
                failure_triage:
                  type: object
                  additionalProperties: true
      responses:
        "201":
          description: Risk report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskReportResponse"
        "400":
          description: Invalid JSON payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal risk evaluation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /failures/triage:
    post:
      summary: Triage a failure and recommend follow-up actions
      operationId: postFailureTriage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                failure_analysis:
                  type: object
                  additionalProperties: true
                execution_record:
                  $ref: "#/components/schemas/ExecutionRecordV1"
                evidence_manifest:
                  $ref: "#/components/schemas/EvidenceManifestV1"
                report:
                  type: object
                  additionalProperties: true
      responses:
        "201":
          description: Failure triage result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailureTriageResponse"
        "400":
          description: Invalid JSON payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal failure triage failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /healing/preview:
    post:
      summary: Preview self-healing advice without modifying YAML or page objects
      operationId: postHealingPreview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelfHealingPreviewRequest"
      responses:
        "200":
          description: Self-healing advice preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SelfHealingPreviewResponse"
        "400":
          description: Invalid JSON payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Route not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal self-healing preview failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /assets/page-objects:
    post:
      summary: Create a page object asset
      operationId: postPageObject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePageObjectRequest"
      responses:
        "201":
          description: Page object created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageObjectAssetResponse"
        "422":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal asset management failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /assets/page-objects/{page}/elements:
    post:
      summary: Add or update an element inside a page object
      operationId: postPageObjectElement
      parameters:
        - in: path
          name: page
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPageElementRequest"
      responses:
        "201":
          description: Page object element added or updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageObjectAssetResponse"
        "422":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal asset management failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /assets/test-cases/sync:
    post:
      summary: Rewrite an existing YAML test case with standard steps inferred from its page object
      operationId: postSyncTestCase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncTestCaseRequest"
      responses:
        "200":
          description: Test case synchronized successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCaseAssetResponse"
        "422":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal asset management failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    RunnerCatalogResponse:
      type: object
      additionalProperties: false
      required:
        - items
        - default_runner
      properties:
        default_runner:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/RunnerProfile"
...
    ExecutionReportResponse:
      type: object
      additionalProperties: false
      required:
        - report
        - execution_record
        - report_json_path
        - report_markdown_path
        - report_summary_path
        - report_summary_preview
      properties:
        report:
          $ref: "#/components/schemas/ExecutionReport"
        execution_record:
          $ref: "#/components/schemas/ExecutionRecordV1"
        report_json_path:
          type: string
        report_markdown_path:
          type: string
        report_summary_path:
          type: string
        report_summary_preview:
          $ref: "#/components/schemas/ReportSummaryPreview"
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - invalid_json
                - not_found
                - unsupported_media_type
                - validation_error
                - runner_failed
                - internal_error
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    PageObjectAssetResponse:
      type: object
      additionalProperties: false
      required:
        - page_object
        - path
      properties:
        page_object:
          type: object
        path:
          type: string
...
    ScaffoldTemplateListResponse:
      type: object
      additionalProperties: false
      required:
        - templates
      properties:
        templates:
          type: array
          items:
            $ref: "#/components/schemas/ScaffoldTemplate"
    ScaffoldTemplateResponse:
      type: object
      additionalProperties: false
      required:
        - template
      properties:
        template:
          $ref: "#/components/schemas/ScaffoldTemplate"